Move new classes and interface to separate package
