+++ src/main/java/com/cloudbees/hudson/plugins/folder/properties/AuthorizationMatrixProperty.java
@@ -28,14 +28,14 @@ import com.cloudbees.hudson.plugins.folder.AbstractFolderProperty;
 import com.cloudbees.hudson.plugins.folder.AbstractFolderPropertyDescriptor;
 import hudson.Extension;
 import hudson.model.Item;
-import hudson.security.AuthorizationMatrixPropertyDescriptor;
+import org.jenkinsci.plugins.matrixauth.AuthorizationMatrixPropertyDescriptor;
 import hudson.security.GlobalMatrixAuthorizationStrategy;
 import hudson.security.Permission;
 import hudson.security.PermissionScope;
 import hudson.security.ProjectMatrixAuthorizationStrategy;
 import hudson.security.SidACL;
-import hudson.security.AbstractMatrixPropertyConverter;
-import hudson.security.AuthorizationProperty;
+import org.jenkinsci.plugins.matrixauth.AbstractMatrixPropertyConverter;
+import org.jenkinsci.plugins.matrixauth.AuthorizationProperty;
 import hudson.util.FormValidation;
 import net.sf.json.JSONObject;
 import org.acegisecurity.acls.sid.Sid;
