+++ src/main/java/hudson/security/AuthorizationMatrixProperty.java
@@ -46,6 +46,9 @@ import javax.servlet.ServletException;
 import net.sf.json.JSONObject;
 import org.acegisecurity.acls.sid.PrincipalSid;
 import org.acegisecurity.acls.sid.Sid;
+import org.jenkinsci.plugins.matrixauth.AbstractMatrixPropertyConverter;
+import org.jenkinsci.plugins.matrixauth.AuthorizationMatrixPropertyDescriptor;
+import org.jenkinsci.plugins.matrixauth.AuthorizationProperty;
 import org.kohsuke.stapler.AncestorInPath;
 import org.kohsuke.stapler.QueryParameter;
 import org.kohsuke.stapler.StaplerRequest;
